vk: Implement depth scaling using hardware blit/copy engines - Removes the old depth scaling using an overlay.   It was never going to work properly due to per-pixel stencil writes being unavailable - TODO: Preserve stencil buffer during ARGB8->D32S8 shader conversion pass

+++ rpcs3/Emu/RSX/rsx_cache.h
@@ -502,7 +502,7 @@ namespace rsx
 				entries.push_back(tmp);
 			}
 
-			if ((entry_count = entries.size()) <= 2)
+			if ((entry_count = (u32)entries.size()) <= 2)
 				return;
 
 			root.rewind();
